Fix checksum tests on 32-bit systems.
Conflicts:

	CHANGES.rst
	lib/pyfits/hdu/base.py

git-svn-id: http://svn6.assembla.com/svn/pyfits/trunk@2784 ed100bfc-0583-0410-97f2-c26b58777a21

+++ lib/pyfits/hdu/base.py
@@ -1541,8 +1541,8 @@ class _ValidHDU(_BaseHDU, _Verify):
 
         hi = sum32 >> u16
         lo = sum32 & uFFFF
-        hi += np.add.reduce(data[0::2])
-        lo += np.add.reduce(data[1::2])
+        hi += np.add.reduce(data[0::2], dtype=np.uint64)
+        lo += np.add.reduce(data[1::2], dtype=np.uint64)
 
         if (data.nbytes // 2) % 2:
             lo += last << u8
